requirements.txt
backcall==0.2.0
backports.zoneinfo==0.2.1
colorama==0.4.4
dateparser==1.1.1
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
ipykernel==6.13.0
ipython==7.32.0
jedi==0.18.1
jupyter-client==7.2.2
jupyter-core==4.10.0
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.21.6
packaging==21.3
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.29
psutil==5.9.0
Pygments==2.11.2
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pyzmq==22.3.0
regex==2022.3.2
six==1.16.0
tornado==6.1
traitlets==5.1.1
tzdata==2022.1
tzlocal==4.2
wcwidth==0.2.5
